requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
exceptiongroup==1.2.0
idna==3.6
iniconfig==2.0.0
packaging==23.2
pluggy==1.4.0
pytest==8.0.2
requests==2.31.0
tomli==2.0.1
urllib3==2.2.1
